Updated the rebase parser test to use a minimal test case rather than the massive dump. Also added a json version of the minimal rebase file to benchmark against.
TODO: Need the community to verify if the minimal is representative (otherwise will need to extend this)

Modified Base structure to have more detailed structure to capture commercial availabilities of the emzymes as a pair of organization and last updated availability.

Stripped the commercial availability parser to get rid of the ridiculous line skip steps (since the new line reader would cause problems with that). Additionally added a regex based line reader for the commercial availability.