feat: add seqfold
Add seqfold package, Go port of the python project at:
https://github.com/Lattice-Automation/seqfold

Add general useful types and maps of energies used to calculate the
folding structures.

Add to checks/checks.go a couple of function to determine if a sequence
is DNA or RNA. This is need by `seqfold` to discriminate for example the
energies used in the folding.