fold: add fold.Result, un-export most of lib
Add `fold.Resutl` to hold the result of fold.Fold and move there as
methods `MinimumFreeEnergy()` and `DotBracket()`.
In this way we can keep the API surface small without compromising the
package functionality, more methods can be added to `fold.Result` in
case we need them, e.g. `result.JSON()` to obtain a representation of the
resulting structure in JSON or other formats.